backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
opencv-python
supervision
numpy
onnxruntime-qnn